requirements.txt
coloredlogs==15.0.1
fastapi==0.90.0
ffmpeg==1.4
numpy==1.24.1
pika==1.3.1
pydantic==1.10.4
pydub==0.25.1
python-dotenv==0.21.1
requests==2.28.2
sox==1.4.1
torch==1.13.0
whisper==1.1.10
